failing build, but all functions moved into cpp
