+++ tools/clang/unittests/HLSL/CMakeLists.txt
@@ -46,6 +46,7 @@ add_clang_library(clang-hlsl-tests SHARED
   PixTest.cpp
   RewriterTest.cpp
   ShaderOpTest.cpp
+  HLSLTestUtils.cpp
   SystemValueTest.cpp
   ValidationTest.cpp
   VerifierTest.cpp
@@ -59,6 +60,7 @@ set(HLSL_IGNORE_SOURCES
   PixTest.cpp
   RewriterTest.cpp
   ShaderOpTest.cpp
+  HLSLTestUtils.cpp
   )
 
 add_clang_unittest(clang-hlsl-tests
